requirements.txt
# requirements.txt

discord.py
Flask